refactor(mergeMapTo): move support types under mergeMapTo.ts
