--- src/operator/mergeMapTo-support.ts
@@ -1,92 +1,0 @@
-import {Operator} from '../Operator';
-import {Observer} from '../Observer';
-import {Observable} from '../Observable';
-import {Subscriber} from '../Subscriber';
-import {tryCatch} from '../util/tryCatch';
-import {errorObject} from '../util/errorObject';
-import {Subscription} from '../Subscription';
-import {OuterSubscriber} from '../OuterSubscriber';
-import {subscribeToResult} from '../util/subscribeToResult';
-
-export class MergeMapToOperator<T, R, R2> implements Operator<Observable<T>, R2> {
-  constructor(private ish: Observable<R> | Promise<R>,
-              private resultSelector?: (outerValue: T, innerValue: R, outerIndex: number, innerIndex: number) => R2,
-              private concurrent: number = Number.POSITIVE_INFINITY) {
-    }
-
-  call(observer: Subscriber<R2>): Subscriber<T> {
-    return new MergeMapToSubscriber(observer, this.ish, this.resultSelector, this.concurrent);
-  }
-}
-
-export class MergeMapToSubscriber<T, R, R2> extends OuterSubscriber<T, R> {
-  private hasCompleted: boolean = false;
-  private buffer: Observable<any>[] = [];
-  private active: number = 0;
-  protected index: number = 0;
-
-  constructor(destination: Subscriber<R2>,
-              private ish: Observable<R> | Promise<R>,
-              private resultSelector?: (outerValue: T, innerValue: R, outerIndex: number, innerIndex: number) => R2,
-              private concurrent: number = Number.POSITIVE_INFINITY) {
-    super(destination);
-  }
-
-  protected _next(value: any): void {
-    if (this.active < this.concurrent) {
-      const resultSelector = this.resultSelector;
-      const index = this.index++;
-      const ish = this.ish;
-      const destination = this.destination;
-
-      this.active++;
-      this._innerSub(ish, destination, resultSelector, value, index);
-    } else {
-      this.buffer.push(value);
-    }
-  }
-
-  private _innerSub(ish: any,
-                    destination: Observer<R>,
-                    resultSelector: (outerValue: T, innerValue: R, outerIndex: number, innerIndex: number) => R2,
-                    value: T,
-                    index: number): void {
-    this.add(subscribeToResult<T, R>(this, ish, value, index));
-  }
-
-  protected _complete(): void {
-    this.hasCompleted = true;
-    if (this.active === 0 && this.buffer.length === 0) {
-      this.destination.complete();
-    }
-  }
-
-  notifyNext(outerValue: T, innerValue: R, outerIndex: number, innerIndex: number): void {
-    const { resultSelector, destination } = this;
-    if (resultSelector) {
-      const result = tryCatch(resultSelector)(outerValue, innerValue, outerIndex, innerIndex);
-      if (result === errorObject) {
-        destination.error(errorObject.e);
-      } else {
-        destination.next(result);
-      }
-    } else {
-      destination.next(innerValue);
-    }
-  }
-
-  notifyError(err: any): void {
-    this.destination.error(err);
-  }
-
-  notifyComplete(innerSub: Subscription): void {
-    const buffer = this.buffer;
-    this.remove(innerSub);
-    this.active--;
-    if (buffer.length > 0) {
-      this._next(buffer.shift());
-    } else if (this.active === 0 && this.hasCompleted) {
-      this.destination.complete();
-    }
-  }
-}
